Factor out FlowNodeGraph
