+++ site/package.json
@@ -11,6 +11,7 @@
   },
   "dependencies": {
     "@ts-ast-viewer/shared": "^1.0.0",
+    "@viz-js/viz": "^3.3.1",
     "circular-json": "^0.5.9",
     "lz-string": "^1.4.4",
     "monaco-editor": "^0.37.1",
